@@ -24,7 +24,7 @@ lint.select = [
     "I", # import sorting
     "U", # upgrade
 ]
-target-version = "py37"
+target-version = "py39"
 line-length = 120
 
 [lint.per-file-ignores]
